OGM-887 Move DotPatternMapHelpers to org.hibernate.ogm.datastore.document.impl
